consolidate formatting functions
