+++ src/lib/formatters.ts
@@ -1,19 +1,45 @@
 import { toZonedTime, format } from 'date-fns-tz';
-interface EventDates {
+
+interface McplEvent {
+  title: string;
+  event_start: string;
+  event_end: string;
+  description: string;
+  url: string;
+}
+
+interface BloomEventDates {
   eventDate: string;
 }
 
-interface Event {
+interface VisitBloomEvent {
   title: string;
   startTime: string;
   endTime: string;
   description: string;
-  dates: EventDates;
+  dates: BloomEventDates;
   color?: string;
   url: string;
 }
 
-export const visitBloomFormatter = (data: Event[]) => {
+const normalizeUrl = (url: string) => {
+  return url.replace(/([^:])\/{2,}/g, '$1/');
+};
+
+export const mcplFormatter = (data: McplEvent[]) => {
+  console.log(data);
+  return data.map((event) => {
+    return {
+      title: event.title,
+      start: event.event_start,
+      end: event.event_end,
+      description: event.description,
+      url: normalizeUrl(event.url),
+    };
+  });
+};
+
+export const visitBloomFormatter = (data: VisitBloomEvent[]) => {
   const timeZone = 'America/Indiana/Indianapolis';
   const baseURL = 'https://www.visitbloomington.com';
   return data.map((event) => {
